+++ variables.tf
@@ -25,6 +25,13 @@ variable "tags" {
 /* -------------------------------------------------------------------------- */
 /*                              CloudWatch Event                              */
 /* -------------------------------------------------------------------------- */
+/* -------------------------------- Event Bus ------------------------------- */
+variable "bus_name" {
+  description = "A unique name for your EventBridge Bus"
+  type        = string
+  default     = "default"
+}
+
 /* ----------------------------- CloudWatch Rule ---------------------------- */
 variable "cloudwatch_event_rule_is_enabled" {
   type        = bool
@@ -45,6 +52,12 @@ variable "schedule_expression" {
 }
 
 /* ---------------------------- CLoudWatch Target --------------------------- */
+variable "cloudwatch_event_target" {
+  description = "A map of objects with EventBridge Target definitions."
+  type        = any
+  default     = {}
+}
+
 variable "cloudwatch_event_target_id" {
   description = "The unique target assignment ID. If missing, will generate a random, unique id."
   type        = string
@@ -56,8 +69,20 @@ variable "cloudwatch_event_target_arn" {
   type        = string
 }
 
+variable "role_arn" {
+  description = "(Optional) The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. Required if ecs_target is used or target in arn is EC2 instance, Kinesis data stream, Step Functions state machine, or Event Bus in different account or region."
+  type        = string
+  default     = ""
+}
+
 variable "input" {
   description = "Valid JSON text passed to the target. Conflicts with input_path and input_transformer."
+  type        = string
+  default     = null
+}
+
+variable "input_path" {
+  description = "(Optional) The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. Conflicts with input and input_transformer."
   type        = string
   default     = null
 }
@@ -81,12 +106,6 @@ variable "input_transformer" {
   default = null
 }
 
-variable "role_arn" {
-  description = "(Optional) The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. Required if ecs_target is used or target in arn is EC2 instance, Kinesis data stream, Step Functions state machine, or Event Bus in different account or region."
-  type        = string
-  default     = ""
-}
-
 variable "event_pattern" {
   description = "(Optional) The event pattern described a JSON object. At least one of schedule_expression or event_pattern is required."
   type        = string
@@ -94,6 +113,7 @@ variable "event_pattern" {
 }
 
 variable "retry_policy" {
+  description = "(Optional) Parameters used when you are providing retry policies. Documented below. A maximum of 1 are allowed."
   type = object({
     maximum_retry_attempts       = number
     maximum_event_age_in_seconds = number
@@ -105,7 +125,16 @@ variable "retry_policy" {
 }
 
 variable "dead_letter_config_arn" {
-  description = "ARN of the SQS queue specified as the target for the dead-letter queue."
+  description = "(Optional) Parameters used when you are providing a dead letter config. Documented below. A maximum of 1 are allowed."
   type        = string
   default     = null
+}
+
+variable "run_command_targets" {
+  description = "(Optional) Parameters used when you are using the rule to invoke Amazon EC2 Run Command. Documented below. A maximum of 5 are allowed."
+  type = list(object({
+    key    = string
+    values = list(string)
+  }))
+  default = []
 }